Se refactoriza el logueo de Firebase Analytics para usar tener una version productiva y otra para testear
