+++ app/src/main/java/ar/edu/unq/pdes/myprivateblog/screens/post_edit/PostEditViewModel.kt
@@ -4,12 +4,12 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import ar.edu.unq.pdes.myprivateblog.data.BlogEntry
 import ar.edu.unq.pdes.myprivateblog.data.EntityID
-import ar.edu.unq.pdes.myprivateblog.logger.AnalyticsLogger
-import ar.edu.unq.pdes.myprivateblog.logger.TypeEventAnalytics
+import ar.edu.unq.pdes.myprivateblog.services.AnalyticsService
 import ar.edu.unq.pdes.myprivateblog.services.PostService
+import ar.edu.unq.pdes.myprivateblog.services.TypeEventAnalytics
 import javax.inject.Inject
 
-class PostEditViewModel @Inject constructor(private val postService: PostService, val analytics: AnalyticsLogger) : ViewModel() {
+class PostEditViewModel @Inject constructor(private val postService: PostService, val analytics: AnalyticsService) : ViewModel() {
 
     enum class State {
         EDITING, SUCCESS, ERROR
